Edit installers to support python outside of virtualenv
The current method of installation was not tested outside of
a virtual environment. The bin location returned by sys.prefix
is different if one is not working in a virtualenv. this will
make the arkouda_server available in the PYTHONPATH if one
has installed it in a virtualenv and has that active or if
they have installed it in their active python distribution.